requirements.txt
# Core RAG
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
chromadb>=0.4.22

# Embeddings
openai>=1.12.0
tiktoken>=0.5.2

# Document Processing
pypdf>=4.0.0
python-docx>=1.1.0
unstructured>=0.12.0

# API
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Observability
structlog>=24.1.0
